pyproject.toml
[project]
name = "whisper-transcription"
version = "0.1.0"
description = "Simple POC for local Whisper speech-to-text transcription"
requires-python = ">=3.8"
dependencies = [
    "openai-whisper>=20231117",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "numpy>=1.24.0",
    "pynput>=1.7.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
